+++ DentalOfficeManagement/src/main/java/com/utn/DentalOfficeManagement/DTO/Response/PlanResponseDTO.java
@@ -9,8 +9,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class PlanResponseDTO {
 
-    private Integer idPlan;
-    private Integer obraSocialId;
+    private Long idPlan;
+    private Long obraSocialId;
     private String nombre;
     private String descripcion;
 }
